requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
bcrypt
python-multipart